# doc/api.yaml
openapi: 3.0.9
info:
  title: WASAText
  version: "0.1"
  description: |
    Simple messaging app API for the WASA project.
    "Login" is simplified and used only for testing (no real security).
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
  /session:
    post:
      tags: [Session]
      summary: Simplified login (create or get user)
      description: |
        Create a user if it does not exist, otherwise reuse the existing user.
        Returns an identifier that must be sent as "Authorization: Bearer <identifier>".
      operationId: doLogin
      security: []
      requestBody:
        required: true
        description: Username to create or reuse.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "201":
          description: User created or reused successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid username.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/name:
    put:
      tags: [Me]
      summary: Change my username
      description: |
        Change the current user's name. The new name must not already be used by another user.
      operationId: setMyUserName
      requestBody:
        required: true
        description: New username.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetNameRequest"
      responses:
        "204":
          description: Username updated successfully.
        "409":
          description: Name already used by another user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/photo:
    put:
      tags: [Me]
      summary: Set my profile photo
      description: |
        Set the current user's profile photo. Use empty string to remove.
      operationId: setMyPhoto
      requestBody:
        required: true
        description: Photo in base64 data URL format (or empty to remove).
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPhotoRequest"
      responses:
        "204":
          description: Photo updated successfully.
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users:
    get:
      tags: [Users]
      summary: List all users
      description: |
        Return all existing WASAText usernames. Used by the UI to start new conversations.
      operationId: listUsers
      responses:
        "200":
          description: List of usernames.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersList"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /conversations:
    get:
      tags: [Conversations]
      summary: List my conversations
      description: |
        Return all conversations the current user belongs to.
        Items include title, isGroup and latest-message preview metadata.
      operationId: getMyConversations
      responses:
        "200":
          description: Conversations list (sorted by latest activity).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversations"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [Conversations]
      summary: Start a direct conversation
      description: |
        Create or return an existing direct conversation with the specified user.
      operationId: createDirectConversation
      requestBody:
        required: true
        description: Target username.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDirectRequest"
      responses:
        "201":
          description: Direct conversation created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationIdResponse"
        "404":
          description: Target user not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /groups:
    post:
      tags: [Groups]
      summary: Create a group conversation
      description: |
        Create a new group conversation with a name and an initial list of members.
        The current user is always included as a member.
      operationId: createGroup
      requestBody:
        required: true
        description: Group name and members.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationIdResponse"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /conversations/{conversationId}:
    get:
      tags: [Conversations]
      summary: Get one conversation (with messages)
      description: |
        Return full conversation data, including messages.
        Messages are returned in reverse chronological order by the backend.
      operationId: getConversation
      parameters:
        - name: conversationId
          in: path
          required: true
          description: Conversation identifier.
          schema:
            $ref: "#/components/schemas/IdString"
      responses:
        "200":
          description: Conversation details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Conversation not found or user is not a member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /conversations/{conversationId}/messages:
    post:
      tags: [Messages]
      summary: Send a message
      description: |
        Send a text or image message to a conversation.
        For text messages "text" is required; for image messages "media" is required.
      operationId: sendMessage
      parameters:
        - name: conversationId
          in: path
          required: true
          description: Conversation identifier.
          schema:
            $ref: "#/components/schemas/IdString"
      requestBody:
        required: true
        description: Message payload.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Invalid message (wrong type, missing text/media, invalid replyTo).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Conversation not found or user is not a member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /messages/{messageId}:
    delete:
      tags: [Messages]
      summary: Delete my message
      description: |
        Delete a message sent by the current user.
      operationId: deleteMessage
      parameters:
        - name: messageId
          in: path
          required: true
          description: Message identifier.
          schema:
            $ref: "#/components/schemas/IdString"
      responses:
        "204":
          description: Message deleted successfully.
        "403":
          description: Current user is not the sender.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Message not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /messages/{messageId}/forward:
    post:
      tags: [Messages]
      summary: Forward message
      description: |
        Forward an existing message to another conversation where the user is a member.
      operationId: forwardMessage
      parameters:
        - name: messageId
          in: path
          required: true
          description: Message identifier.
          schema:
            $ref: "#/components/schemas/IdString"
      requestBody:
        required: true
        description: Target conversation for forwarding.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForwardRequest"
      responses:
        "201":
          description: Forwarded message created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Message or conversation not found (or user not member).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /messages/{messageId}/comments:
    post:
      tags: [Messages]
      summary: React to a message
      description: |
        Add (or update) the current user's reaction to a message with an emoji.
        A user can have at most one reaction per message.
      operationId: commentMessage
      parameters:
        - name: messageId
          in: path
          required: true
          description: Message identifier.
          schema:
            $ref: "#/components/schemas/IdString"
      requestBody:
        required: true
        description: Emoji reaction.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactRequest"
      responses:
        "201":
          description: Reaction created/updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionIdResponse"
        "404":
          description: Message not found or user not member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /messages/{messageId}/comments/{reactionId}:
    delete:
      tags: [Messages]
      summary: Remove reaction
      description: |
        Remove a reaction previously created by the current user.
      operationId: uncommentMessage
      parameters:
        - name: messageId
          in: path
          required: true
          description: Message identifier.
          schema:
            $ref: "#/components/schemas/IdString"
        - name: reactionId
          in: path
          required: true
          description: Reaction identifier.
          schema:
            $ref: "#/components/schemas/IdString"
      responses:
        "204":
          description: Reaction deleted successfully.
        "403":
          description: Current user is not the author of the reaction.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Message or reaction not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /groups/{groupId}/name:
    put:
      tags: [Groups]
      summary: Set group name
      description: |
        Change group title. Only members can change group data.
      operationId: setGroupName
      parameters:
        - name: groupId
          in: path
          required: true
          description: Group identifier (same as conversationId for group conversations).
          schema:
            $ref: "#/components/schemas/IdString"
      requestBody:
        required: true
        description: New group name.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetGroupNameRequest"
      responses:
        "204":
          description: Group name updated successfully.
        "404":
          description: Group not found or user not member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /groups/{groupId}/photo:
    put:
      tags: [Groups]
      summary: Set group photo
      description: |
        Update group photo. Use empty string to remove.
      operationId: setGroupPhoto
      parameters:
        - name: groupId
          in: path
          required: true
          description: Group identifier (same as conversationId for group conversations).
          schema:
            $ref: "#/components/schemas/IdString"
      requestBody:
        required: true
        description: Photo in base64 data URL format (or empty to remove).
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPhotoRequest"
      responses:
        "204":
          description: Group photo updated successfully.
        "404":
          description: Group not found or user not member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /groups/{groupId}/members:
    post:
      tags: [Groups]
      summary: Add user to group
      description: |
        Add an existing user to the group. Only members can add.
      operationId: addToGroup
      parameters:
        - name: groupId
          in: path
          required: true
          description: Group identifier.
          schema:
            $ref: "#/components/schemas/IdString"
      requestBody:
        required: true
        description: Username to add.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToGroupRequest"
      responses:
        "204":
          description: User added successfully.
        "404":
          description: Group or user not found (or user not member).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /groups/{groupId}/leave:
    post:
      tags: [Groups]
      summary: Leave group
      description: |
        Remove the current user from the group.
      operationId: leaveGroup
      parameters:
        - name: groupId
          in: path
          required: true
          description: Group identifier.
          schema:
            $ref: "#/components/schemas/IdString"
      responses:
        "204":
          description: Left group successfully.
        "404":
          description: Group not found or user not member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not logged (missing/invalid bearer identifier).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Simplified bearer identifier used only for testing/impersonation.
  schemas:
...
    IdString:
      type: string
      description: Generic identifier used by the API.
      minLength: 1
      maxLength: 64
      pattern: "^[A-Za-z0-9_\\-]+$"
...
    Error:
      type: object
      description: Error response with a human-readable message.
      properties:
        message:
          type: string
          description: Human-readable error message.
          minLength: 1
          maxLength: 200
          pattern: "^.+$"
      required: [message]

    # ---- Login
    LoginRequest:
      type: object
      description: Request body for simplified login.
      properties:
        name:
          $ref: "#/components/schemas/Username"
      required: [name]
    LoginResponse:
      type: object
      description: Response of simplified login.
      properties:
        identifier:
          type: string
          description: Bearer identifier to use in Authorization header.
          minLength: 1
          maxLength: 128
          pattern: "^.+$"
      required: [identifier]

    # ---- Profile updates
    SetNameRequest:
      type: object
      description: Request body for changing username.
      properties:
        name:
          $ref: "#/components/schemas/Username"
      required: [name]
    SetPhotoRequest:
      type: object
      description: Request body for setting/removing a photo.
      properties:
        photo:
          $ref: "#/components/schemas/Base64DataUrl"
      required: [photo]

    # ---- Lists
    UsersList:
      type: object
      description: Response containing all usernames.
      properties:
        users:
          type: array
          description: List of usernames.
          minItems: 0
          maxItems: 10000
          items:
            $ref: "#/components/schemas/Username"
      required: [users]

    # ---- Conversations
...
    Conversations:
      type: object
      description: List of conversations for the current user.
      properties:
        conversations:
          type: array
          description: Conversations sorted by latest activity.
          minItems: 0
          maxItems: 10000
          items:
            $ref: "#/components/schemas/ConversationItem"
      required: [conversations]
    Conversation:
      type: object
      description: Full conversation including messages.
      properties:
        id:
          $ref: "#/components/schemas/IdString"
        title:
          type: string
          description: Conversation title.
          minLength: 1
          maxLength: 64
          pattern: "^.+$"
        isGroup:
          type: boolean
          description: True if group conversation.
        members:
          type: array
          description: Usernames of members.
          minItems: 1
          maxItems: 500
          items:
            $ref: "#/components/schemas/Username"
        photo:
          $ref: "#/components/schemas/Base64DataUrl"
        messages:
          type: array
          description: Messages (returned newest-first by backend).
          minItems: 0
          maxItems: 20000
          items:
            $ref: "#/components/schemas/Message"
      required: [id, title, isGroup, members, messages]

    # ---- Requests
    CreateDirectRequest:
      type: object
      description: Request body to create a direct conversation.
      properties:
        username:
          $ref: "#/components/schemas/Username"
      required: [username]
    CreateGroupRequest:
      type: object
      description: Request body to create a group conversation.
      properties:
        name:
          type: string
          description: Group name.
          minLength: 1
          maxLength: 64
          pattern: "^.+$"
        members:
          type: array
          description: Initial members to add.
          minItems: 1
          maxItems: 500
          items:
            $ref: "#/components/schemas/Username"
      required: [name, members]
    SendMessageRequest:
      type: object
      description: Request body to send a message.
      properties:
        type:
          type: string
          description: Message type.
          minLength: 4
          maxLength: 5
          pattern: "^(text|image)$"
          enum: [text, image]
        text:
          type: string
          description: Text content (required if type=text).
          minLength: 0
          maxLength: 5000
          pattern: "^[\\s\\S]*$"
        media:
          $ref: "#/components/schemas/Base64DataUrl"
        replyTo:
          $ref: "#/components/schemas/IdString"
      required: [type]
...
    ConversationIdResponse:
      type: object
      description: Response containing a created conversation identifier.
      properties:
        conversationId:
          $ref: "#/components/schemas/IdString"
      required: [conversationId]
...